Move fully refactored code into new environment subpackage
Eventually, all environment code (including the stuff in the current handler
subpackage) will live in this subpackage.
