+++ fvm/environment/program_logger.go
@@ -1,4 +1,4 @@
-package fvm
+package environment
 
 import (
 	"time"
