+++ fvm/environment/unsafe_random_generator.go
@@ -1,4 +1,4 @@
-package fvm
+package environment
 
 import (
 	"encoding/binary"
